Co-locate the player sheets with the bottom accessory in one modifier
The matched zoom warning persisted because the sheet (in `MainTabView`)
and the matched source (inside the bottom accessory) lived in different
view contexts, so the source view's lifecycle couldn't be tied to the
sheet's transition lifecycle. Mirror the working pattern from another
project: declare a local `@Namespace` inside the modifier and host both
the `tabViewBottomAccessory` and the two `.sheet` modifiers in the same
scope. The sheet's `navigationTransition(.zoom(...))` now references the
exact same namespace as the source view's `matchedTransitionSource`, and
they share the same view-tree update cycle.

Also split the mini player views into bar-only views (`MiniPlayerBar`,
`YouTubeMiniPlayerBar`); the surrounding Button is now provided by the
accessory modifier so the matched source is applied directly to the
button label content.
